引力模型初步完善，问题：为何出现跳点现象？

+++ src/initiator/Parser.java
@@ -139,7 +139,7 @@ public class Parser {
 		String life = "1m"; // 生存时间默认1分钟
 		String name = ""; // Message的名字
 		int priority = 5; // 优先级
-		int second = 2; // Message将于多长时间后发送
+		double second = 2.0; // Message将于多长时间后发送
 		int share = 1; // 可接收此Message的instance的最大数量
 		
 		List<MessageContent> mc = new ArrayList<MessageContent>();
@@ -161,7 +161,7 @@ public class Parser {
 			} else if (messageA.getName().equals("priority")) {
 				priority = Integer.parseInt(messageA.getValue());
 			} else if (messageA.getName().equals("second")) {
-				second = Integer.parseInt(messageA.getValue());
+				second = Double.parseDouble(messageA.getValue());
 			} else { // share属性
 				share = Integer.parseInt(messageA.getValue());
 			}
